David final

+++ src/main/java/org/uv/tpcs_practica01/Empleado.java
@@ -6,7 +6,7 @@ public class Empleado {
         return clave;
     }
 
-    public void setClave(int clave) {
+    public void setClave(long clave) {
         this.clave = clave;
     }
